Speed Check is back, KI KP and KD are defined, PID early PWM Min added
