+++ Code/lib/openGLCD/config/ks0108/PinConfig_ks0108-Mega.h
@@ -32,30 +32,30 @@
  * Data pin definitions
  * This version uses pins 22-29 for LCD Data 
  */
-#define glcdPinData0    44
-#define glcdPinData1    26
-#define glcdPinData2    42
-#define glcdPinData3    28
-#define glcdPinData4    40
-#define glcdPinData5    30
-#define glcdPinData6    38
-#define glcdPinData7    32
+#define glcdPinData0 44
+#define glcdPinData1 26
+#define glcdPinData2 42
+#define glcdPinData3 28
+#define glcdPinData4 40
+#define glcdPinData5 30
+#define glcdPinData6 38
+#define glcdPinData7 32
 
 /*
  * Pins used for Control
  */
-#define glcdPinCSEL1     36    // CS1 Bit  
-#define glcdPinCSEL2     34    // CS2 Bit
-#define glcdPinRW        46    // R/W Bit
-#define glcdPinDI        22    // D/I Bit 
-#define glcdPinEN        24    // EN Bit
+#define glcdPinCSEL1 36 // CS1 Bit
+#define glcdPinCSEL2 34 // CS2 Bit
+#define glcdPinRW 46    // R/W Bit
+#define glcdPinDI 22    // D/I Bit
+#define glcdPinEN 24    // EN Bit
 
 #if NBR_CHIP_SELECT_PINS > 2
-#define glcdPinCSEL3     32   // third chip select if needed
+#define glcdPinCSEL3 32 // third chip select if needed
 #endif
 
 #if NBR_CHIP_SELECT_PINS > 3
-#define glcdPinCSEL4     31   // fourth chip select if needed
+#define glcdPinCSEL4 31 // fourth chip select if needed
 #endif
 
 // Reset - uncomment the next line if glcd module reset is connected to an Arduino pin
