+++ Code/src/PID.cpp
@@ -1,18 +1,20 @@
 
 #include <PID.h>
 
-PID::PID(float KP, float KI, float KD, int ignorePIDCount){
+PID::PID(float KP, float KI, float KD, int initialPWM, int ignorePIDCount, int stepGaurd){
     this->KP=KP;
     this->KI=KI;
     this->KD=KD;
-    this->ignorePIDCount = ignorePIDCount;
+    this->initialPWM      = initialPWM;
+    this->ignorePIDCount  = ignorePIDCount;
+    this->stepGaurd       = stepGaurd;
 }
 
 int PID::Calc(float desired, float pv){
     float pwm = 0;
     if(ignoreCounter==0)
     {
-        pwm = 20;
+        pwm    = initialPWM;
         oldPWM = pwm;
     }
     if(ignoreCounter>ignorePIDCount)
@@ -25,15 +27,20 @@ int PID::Calc(float desired, float pv){
 
         errorPre = error;
         
-        //pwm = 255 - (KP*error + KI*integral + KD*derivative);  
-        
-        pwm = KP*error + KI*integral + KD*derivative;  
+        pwm = KP*error + KI*integral + KD*derivative;
 
+        if (stepCounter < stepGaurd){         
+            if(pwm < initialPWM)
+                pwm = initialPWM;
+        }
+        
         realPidVal = pwm ; 
 
         pwm = limitOutput(pwm);
 
         oldPWM = pwm;
+
+        stepCounter++;
     }
     
     ignoreCounter++;
@@ -77,4 +84,5 @@ void PID::resetParams(){
     this->integral=0;
     this->derivative=0;
     this->ignoreCounter=0;
+    this->stepCounter=0;
 }
