+++ Code/src/PID.h
@@ -17,9 +17,12 @@ private:
     float timeStep=5e-3; //Unit is in Seconds
     float oldPWM=0;
     float ignorePIDCount=0;
-    int ignoreCounter=0;
+    int   ignoreCounter=0;
+    int   stepCounter=0;
+    int   stepGaurd=0;
+    int   initialPWM=0;
 public:
-    PID(float KP, float KI, float KD, int ignorePIDCount);
+    PID(float KP, float KI, float KD, int initialPWM, int ignorePIDCount, int stepGaurd);
     int   Calc(float desired, float pv);
     void  setTimeStep(float timeStep);
     float getTimeStep();
